+++ ORStools/common/client.py
@@ -35,6 +35,8 @@ from typing import Union, Dict, List, Optional
 from urllib.parse import urlencode
 
 from qgis.PyQt.QtCore import QObject, pyqtSignal
+from qgis.utils import iface
+from qgis.core import Qgis
 from requests.utils import unquote_unreserved
 
 from ORStools import __version__
@@ -179,6 +181,10 @@ class Client(QObject):
                 # noinspection PyUnresolvedReferences
                 self.overQueryLimit.emit()
                 logger.log(f"{e.__class__.__name__}: {str(e)}", 1)
+
+                iface.messageBar().pushMessage(
+                    "ORSTools", "Rate limit exceeded, retrying...", level=Qgis.Warning, duration=2
+                )
 
                 return self.request(url, params, first_request_time, retry_counter + 1, post_json)